Gui qt

+++ pySimBlocks/blocks/systems/linear_state_space.py
@@ -45,15 +45,16 @@ class LinearStateSpace(Block):
         B: np.ndarray,
         C: np.ndarray,
         x0: np.ndarray | None = None,
-        sample_time:float|None = None):
+        sample_time: float | None = None
+    ):
         super().__init__(name, sample_time)
 
         # ------------------------------------------------------------------
         # Store and check matrices
         # ------------------------------------------------------------------
-        self.A = np.asarray(A)
-        self.B = np.asarray(B)
-        self.C = np.asarray(C)
+        self.A = np.asarray(A, dtype=float)
+        self.B = np.asarray(B, dtype=float)
+        self.C = np.asarray(C, dtype=float)
 
         n = self.A.shape[0]
 
@@ -70,9 +71,9 @@ class LinearStateSpace(Block):
         # Initial state x0
         # ------------------------------------------------------------------
         if x0 is None:
-            x0 = np.zeros((n, 1))
+            x0 = np.zeros((n, 1), dtype=float)
         else:
-            x0 = np.asarray(x0).reshape(-1, 1)
+            x0 = np.asarray(x0, dtype=float).reshape(-1, 1)
             if x0.shape != (n, 1):
                 raise ValueError(f"x0 must have shape ({n}, 1).")
